Separate classes in files
